added code that highlights special single line comments.
specifically single line comments like //*** or //! will show up as red. We can also use this to add more specialized comments in the future if wanted. The code is set up at this point to only require adding a tag in the package.json file to add a new comment style.